backend/requirements.txt
# --- Core Framework ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# --- Data Processing ---
numpy==1.26.4
scipy==1.13.1
pandas==2.2.3
PyYAML==6.0.1
pydantic>=2.0.0

# --- LangChain Ecosystem ---
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-google-genai>=0.0.11
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
sentence-transformers>=2.2.0
huggingface-hub==0.24.7

# --- PDF & OCR Tools ---
unstructured>=0.12.0
pdf2image>=1.16.0
pdfminer.six>=20221105
pikepdf>=8.0.0
pypdf>=3.17.0
openpyxl>=3.1.0

# --- Vector Database ---
chromadb>=0.4.22
